Refactor authentication-related classes to use updated User entity and repository, enhance JWT cookie handling, and improve token management methods
