+++ FeedBloom/src/main/java/sams/feedbloom/authentication/service/CustomUserDetailsService.java
@@ -4,8 +4,8 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-import sams.feedbloom.authentication.entity.User;
-import sams.feedbloom.authentication.repository.UserRepository;
+import sams.feedbloom.user.entity.User;
+import sams.feedbloom.user.repository.UserRepository;
 
 @Service
 public class CustomUserDetailsService implements UserDetailsService {
