+++ FeedBloom/src/main/java/sams/feedbloom/authentication/util/JwtUtil.java
@@ -2,10 +2,15 @@ package sams.feedbloom.authentication.util;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
+import sams.feedbloom.authentication.dto.AuthResponse;
 
 import javax.crypto.SecretKey;
+import java.util.Base64;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
@@ -13,8 +18,27 @@ import java.util.function.Function;
 
 @Component
 public class JwtUtil {
-	private static final long EXPIRATION_TIME = 86400000; // 24 hours
-	private final SecretKey SECRET_KEY = Jwts.SIG.HS256.key().build(); // Secure key
+	
+	private static final long EXPIRATION_TIME = 24 * 60 * 60 * 1000; // 24 hours
+	private static final String JWT_COOKIE_NAME = "jwt";
+	private static final SecretKey SECRET_KEY = Keys.hmacShaKeyFor(
+			Base64.getEncoder().encodeToString("your-secure-secret-key-should-be-long".getBytes()).getBytes());
+	
+	public static void setJwtCookie(HttpServletResponse response, AuthResponse authResponse) {
+		Cookie jwtCookie = new Cookie(JWT_COOKIE_NAME, authResponse.getToken());
+		jwtCookie.setHttpOnly(true);
+		jwtCookie.setPath("/web");
+		jwtCookie.setMaxAge((int) (EXPIRATION_TIME / 1000)); // Convert milliseconds to seconds
+		response.addCookie(jwtCookie);
+	}
+	
+	public static void invalidateJwtCookie(HttpServletResponse response) {
+		Cookie jwtCookie = new Cookie(JWT_COOKIE_NAME, null);
+		jwtCookie.setHttpOnly(true);
+		jwtCookie.setPath("/web");
+		jwtCookie.setMaxAge(0);
+		response.addCookie(jwtCookie);
+	}
 	
 	public String generateToken(UserDetails userDetails) {
 		Map<String, Object> claims = new HashMap<>();
@@ -25,13 +49,13 @@ public class JwtUtil {
 		return Jwts.builder()
 		           .claims(claims)
 		           .subject(email)
-		           .issuedAt(new Date(System.currentTimeMillis()))
+		           .issuedAt(new Date())
 		           .expiration(new Date(System.currentTimeMillis() + EXPIRATION_TIME))
-		           .signWith(SECRET_KEY, Jwts.SIG.HS256)
+		           .signWith(SECRET_KEY)
 		           .compact();
 	}
 	
-	public Boolean validateToken(String token, UserDetails userDetails) {
+	public boolean validateToken(String token, UserDetails userDetails) {
 		final String email = extractEmail(token);
 		return (email.equals(userDetails.getUsername()) && !isTokenExpired(token));
 	}
@@ -57,7 +81,7 @@ public class JwtUtil {
 		           .getPayload();
 	}
 	
-	private Boolean isTokenExpired(String token) {
+	private boolean isTokenExpired(String token) {
 		return extractExpiration(token).before(new Date());
 	}
 }
