+++ FeedBloom/src/main/java/sams/feedbloom/comment/entity/Comment.java
@@ -3,9 +3,9 @@ package sams.feedbloom.comment.entity;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
-import sams.feedbloom.authentication.entity.User;
 import sams.feedbloom.common.entity.Common;
 import sams.feedbloom.feedback.entity.Feedback;
+import sams.feedbloom.user.entity.User;
 
 @Getter
 @Setter
